Github workflow and python update script

+++ python/gym/cpr_gym/__main__.py
@@ -1,8 +1,9 @@
 import argparse
 import cpr_gym
-import urllib.request
+import os
+import requests
 import shutil
-
+import sys
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
@@ -12,11 +13,27 @@ if __name__ == "__main__":
     )
     args = parser.parse_args()
     if args.update:
-        print(f"downloading bridge.so to {cpr_gym.dll_name}_new")
-        urllib.request.urlretrieve(cpr_gym.dll_url, cpr_gym.dll_name + "_new")
-        print(f"moving {cpr_gym.dll_name} -> {cpr_gym.dll_name}_old")
-        shutil.move(cpr_gym.dll_name, cpr_gym.dll_name + "_old")
-        print(f"moving {cpr_gym.dll_name}_new -> {cpr_gym.dll_name}")
-        shutil.move(cpr_gym.dll_name + "_new", cpr_gym.dll_name)
+        rel = requests.get(
+            "https://api.github.com/repos/pkel/cpr/releases/latest"
+        ).json()
+        tag = rel["tag_name"]
+        if sys.platform not in ["linux", "darwin"]:
+            raise OSError(
+                f"Unsupported OS: '{sys.platform}' - we support Linux and MacOS"
+            )
+        repo_url = "https://github.com/pkel/cpr"
+        dll_url = f"{repo_url}/releases/download/{tag}/cpr-{tag}-{sys.platform}.so"
+
+        if os.path.exists(cpr_gym.dll_name):
+            print(f"backup bridge to {os.path.relpath(cpr_gym.dll_name)}_old")
+            shutil.move(cpr_gym.dll_name, cpr_gym.dll_name + "_old")
+
+        print(f"download new bridge from {dll_url}")
+        so = requests.get(dll_url, stream=True)
+        so.raise_for_status()
+        with open(f"{cpr_gym.dll_name}", "wb") as handle:
+            for block in so.iter_content(1024):
+                handle.write(block)
+
     if args.version:
         print(cpr_gym.engine.cpr_lib_version)
